MFL: Change PassState to u16. Can't fit 11 flags in 8 bits... go figure.

+++ crates/mfl/src/pass_manager/mod.rs
@@ -17,7 +17,7 @@ mod passes;
 pub mod static_analysis;
 
 flags! {
-    enum PassState: u8 {
+    enum PassState: u16 {
         IdentResolvedSignature,
         IdentResolvedBody,
         DeclareStructs,
